refactor(cli): split config.rs into a config/ module directory
config.rs (1985 lines) becomes config/{schema,convert,load}.rs with tests
mirrored under config/tests/. schema = the *Config types + defaults; convert
= the to_*/resolve_* runtime conversions; load = load_config/validation/
db-path resolution. mod.rs re-exports all three so crate::config::X paths are
unchanged; main.rs mod config; still resolves to the directory. Cross-file and
test-reached private helpers bumped to pub(crate). No prod logic or test
changed (76 tests, 4 #[serial], all preserved).
