+++ crates/cartog/src/config/mod.rs
@@ -1,0 +1,12 @@
+//! Configuration: schema types, runtime conversions, and file loading.
+
+mod convert;
+mod load;
+mod schema;
+
+pub use convert::*;
+pub use load::*;
+pub use schema::*;
+
+#[cfg(test)]
+mod tests;
