+++ crates/cartog/src/config/tests/mod.rs
@@ -1,0 +1,3 @@
+mod convert;
+mod load;
+mod schema;
